rename Component to Reading
